pyproject.toml
[tool.poetry]
name = "store_demo"
version = "0.1.0"
description = ""
authors = ["Manoeljr <manoelv8@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.31"
databases = "^0.5.5"
asyncpg = "^0.25.0"
psycopg2 = "^2.9.3"
psycopg2-binary = "^2.9.3"
fastapi = "^0.73.0"
uvicorn = "^0.17.5"
alembic = "^1.7.6"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
